change the payload of audit event from json into yaml
